+++ app/(commerce)/_layout.tsx
@@ -9,8 +9,8 @@ export default function CommerceLayout() {
       }}
     >
       <Stack.Screen name="subscriptions" />
-      <Stack.Screen name="tree-planting-sub" />
-      {/* <Stack.Screen name="carbon-credit-sub" /> */}
+      {/* <Stack.Screen name="tree-planting-sub" /> */}
+      <Stack.Screen name="carbon-credit-sub" />
       {/* <Stack.Screen name="newsletter-sub" /> */}
       {/* <Stack.Screen name="subscriptions-settings" /> */}
       <Stack.Screen name="shopping-cart" />
