+++ types/cart.ts
@@ -1,0 +1,16 @@
+interface CartItem {
+  id: string;
+  name: string;
+  productType: string;
+  quantity: number;
+}
+
+interface TransactionItem {
+  id: string;
+  name: string;
+  productType: string;
+  quantity: number;
+  price: number;
+}
+
+export { CartItem, TransactionItem };
